+++ frontend/src/metabase/query_builder/actions/index.ts
@@ -1,0 +1,10 @@
+export * from "./core";
+export * from "./models";
+export * from "./native";
+export * from "./navigation";
+export * from "./object-detail";
+export * from "./querying";
+export * from "./sharing";
+export * from "./timelines";
+export * from "./ui";
+export * from "./visualization-settings";
